requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.1
fastapi==0.116.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.10.0
jose==1.0.0
lxml==5.3.0
multidict==6.6.3
openai==1.93.1
azure-cognitiveservices-speech==1.34.0
passlib==1.7.4
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
celery==5.3.6
redis==5.0.1
celery[redis,rabbitmq]
Pillow==10.4.0
requests==2.32.3
boto3==1.34.69
google-cloud-texttospeech==2.16.2
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
opensearch-py==2.3.0